// ts/gen/buf/knit/demo/swapi/film/v1/film_knit.ts
// Copyright 2023 Buf Technologies, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-knit-ts v0.0.1 with parameter "target=ts"
// @generated from file buf/knit/demo/swapi/film/v1/film.proto (package buf.knit.demo.swapi.film.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

import type { Date } from "../../../../../../google/type/date_knit.js";

export type FilmService = {
  "buf.knit.demo.swapi.film.v1.FilmService": {
    fetch: {
      getFilms: { $: GetFilmsRequest; value: GetFilmsResponse; };
      listFilms: { $: ListFilmsRequest; value: ListFilmsResponse; };
    };
    do: {
    };
    listen: {
    };
  };
};

export interface Film {
  id: string;
  title: string;
  episodeNumber: number;
  openingCrawl: string;
  director: string;
  producer: string;
  releaseDate?: Date;
  characterIds: Array<string>;
  planetIds: Array<string>;
  speciesIds: Array<string>;
  starshipIds: Array<string>;
  vehicleIds: Array<string>;
  created?: "@wkt/Timestamp";
  edited?: "@wkt/Timestamp";
};

export interface GetFilmsRequest {
  ids: Array<string>;
};

export interface GetFilmsResponse {
  films: Array<Film>;
};

export interface ListFilmsRequest {
  pageSize: number;
  pageToken: string;
};

export interface ListFilmsResponse {
  films: Array<Film>;
  nextPageToken: string;
};