// ts/gen/buf/knit/demo/swapi/person/v1/person_knit.ts
// Copyright 2023 Buf Technologies, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-knit-ts v0.0.1 with parameter "target=ts"
// @generated from file buf/knit/demo/swapi/person/v1/person.proto (package buf.knit.demo.swapi.person.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

export type PersonService = {
  "buf.knit.demo.swapi.person.v1.PersonService": {
    fetch: {
      getPeople: { $: GetPeopleRequest; value: GetPeopleResponse; };
      listPeople: { $: ListPeopleRequest; value: ListPeopleResponse; };
    };
    do: {
    };
    listen: {
    };
  };
};

export interface Person {
  id: string;
  name: string;
  mass: string;
  hairColor: string;
  skinColor: string;
  eyeColor: string;
  birthYear: string;
  gender: string;
  homeworldId: string;
  speciesIds: Array<string>;
  filmIds: Array<string>;
  starshipIds: Array<string>;
  vehicleIds: Array<string>;
  created?: "@wkt/Timestamp";
  edited?: "@wkt/Timestamp";
};

export interface GetPeopleRequest {
  ids: Array<string>;
};

export interface GetPeopleResponse {
  people: Array<Person>;
};

export interface ListPeopleRequest {
  pageSize: number;
  pageToken: string;
};

export interface ListPeopleResponse {
  people: Array<Person>;
  nextPageToken: string;
};