// ts/gen/buf/knit/demo/swapi/planet/v1/planet_knit.ts
// Copyright 2023 Buf Technologies, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-knit-ts v0.0.1 with parameter "target=ts"
// @generated from file buf/knit/demo/swapi/planet/v1/planet.proto (package buf.knit.demo.swapi.planet.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

export type PlanetService = {
  "buf.knit.demo.swapi.planet.v1.PlanetService": {
    fetch: {
      getPlanets: { $: GetPlanetsRequest; value: GetPlanetsResponse; };
      listPlanets: { $: ListPlanetsRequest; value: ListPlanetsResponse; };
    };
    do: {
    };
    listen: {
    };
  };
};

export interface Planet {
  id: string;
  climates: Array<string>;
  diameter: number;
  gravity: string;
  name: string;
  orbitalPeriod: number;
  population: number;
  rotationPeriod: number;
  surfaceWater: number;
  terrains: Array<string>;
  residentIds: Array<string>;
  filmIds: Array<string>;
  created?: "@wkt/Timestamp";
  edited?: "@wkt/Timestamp";
};

export interface GetPlanetsRequest {
  ids: Array<string>;
};

export interface GetPlanetsResponse {
  planets: Array<Planet>;
};

export interface ListPlanetsRequest {
  pageSize: number;
  pageToken: string;
};

export interface ListPlanetsResponse {
  planets: Array<Planet>;
  nextPageToken: string;
};