// ts/gen/buf/knit/demo/swapi/species/v1/species_knit.ts
// Copyright 2023 Buf Technologies, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-knit-ts v0.0.1 with parameter "target=ts"
// @generated from file buf/knit/demo/swapi/species/v1/species.proto (package buf.knit.demo.swapi.species.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

export type SpeciesService = {
  "buf.knit.demo.swapi.species.v1.SpeciesService": {
    fetch: {
      getSpecies: { $: GetSpeciesRequest; value: GetSpeciesResponse; };
      listSpecies: { $: ListSpeciesRequest; value: ListSpeciesResponse; };
    };
    do: {
    };
    listen: {
    };
  };
};

export interface Species {
  id: string;
  averageHeight: number;
  averageLifespan: bigint;
  classification: string;
  designation: string;
  eyeColors: Array<string>;
  hairColors: Array<string>;
  language: string;
  name: string;
  skinColors: Array<string>;
  homeworldId: string;
  peopleIds: Array<string>;
  filmIds: Array<string>;
  created?: "@wkt/Timestamp";
  edited?: "@wkt/Timestamp";
};

export interface GetSpeciesRequest {
  ids: Array<string>;
};

export interface GetSpeciesResponse {
  species: Array<Species>;
};

export interface ListSpeciesRequest {
  pageSize: number;
  pageToken: string;
};

export interface ListSpeciesResponse {
  species: Array<Species>;
  nextPageToken: string;
};