// ts/gen/buf/knit/demo/swapi/vehicle/v1/vehicle_knit.ts
// Copyright 2023 Buf Technologies, Inc.
//
// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
//      http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

// @generated by protoc-gen-knit-ts v0.0.1 with parameter "target=ts"
// @generated from file buf/knit/demo/swapi/vehicle/v1/vehicle.proto (package buf.knit.demo.swapi.vehicle.v1, syntax proto3)
/* eslint-disable */
// @ts-nocheck

export type VehicleService = {
  "buf.knit.demo.swapi.vehicle.v1.VehicleService": {
    fetch: {
      getVehicles: { $: GetVehiclesRequest; value: GetVehiclesResponse; };
      listVehicles: { $: ListVehiclesRequest; value: ListVehiclesResponse; };
    };
    do: {
    };
    listen: {
    };
  };
};

export interface Vehicle {
  id: string;
  name: string;
  cargoCapacity: number;
  consumables: string;
  costInCredits?: bigint;
  crew: bigint;
  length: number;
  manufacturers: Array<string>;
  model: string;
  maxAtmospheringSpeed?: bigint;
  passengers: bigint;
  class: string;
  pilotIds: Array<string>;
  filmIds: Array<string>;
  created?: "@wkt/Timestamp";
  edited?: "@wkt/Timestamp";
};

export interface GetVehiclesRequest {
  ids: Array<string>;
};

export interface GetVehiclesResponse {
  vehicles: Array<Vehicle>;
};

export interface ListVehiclesRequest {
  pageSize: number;
  pageToken: string;
};

export interface ListVehiclesResponse {
  vehicles: Array<Vehicle>;
  nextPageToken: string;
};